+++ src/Query/Filter/Fields/RelationalField.php
@@ -1,10 +1,14 @@
-<?php
+<?php /** @noinspection PhpUnused */
 
-namespace Bayfront\Directus\Query\Fields;
+namespace Bayfront\Directus\Query\Filter\Fields;
 
 use Bayfront\ArrayHelpers\Arr;
 use Bayfront\Directus\Query\Interfaces\FieldInterface;
 
+/**
+ * Relational field filter.
+ * See: https://docs.directus.io/reference/filter-rules.html#relational
+ */
 class RelationalField implements FieldInterface
 {