+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added query builder classes.
+- Added query builder.
 
 ## [1.0.0] - 2024.05.07